Move messaging stuff in the same directory in lib-components
