+++ frontend/src/citizen-frontend/messages/MessageEditor.tsx
@@ -26,7 +26,7 @@ import InputField from 'lib-components/atoms/form/InputField'
 import MultiSelect from 'lib-components/atoms/form/MultiSelect'
 import { desktopMin } from 'lib-components/breakpoints'
 import { FixedSpaceFlexWrap } from 'lib-components/layout/flex-helpers'
-import { ToggleableRecipient } from 'lib-components/molecules/ToggleableRecipient'
+import { ToggleableRecipient } from 'lib-components/messages/ToggleableRecipient'
 import { Bold, P } from 'lib-components/typography'
 import { defaultMargins, Gap } from 'lib-components/white-space'
 import colors from 'lib-customizations/common'
