+++ frontend/src/citizen-frontend/messages/ThreadListItem.tsx
@@ -12,14 +12,14 @@ import {
   FixedSpaceColumn,
   FixedSpaceRow
 } from 'lib-components/layout/flex-helpers'
-import FileDownloadButton from 'lib-components/molecules/FileDownloadButton'
 import {
   Container,
   DeleteThreadButton,
   Header,
   TitleAndDate,
   Truncated
-} from 'lib-components/molecules/ThreadListItem'
+} from 'lib-components/messages/ThreadListItem'
+import FileDownloadButton from 'lib-components/molecules/FileDownloadButton'
 
 import { getAttachmentUrl } from '../attachments'
 import { useTranslation } from '../localization'
