+++ frontend/src/employee-frontend/components/Header.tsx
@@ -13,8 +13,8 @@ import { combine } from 'lib-common/api'
 import { EvakaLogo } from 'lib-components/atoms/EvakaLogo'
 import InlineButton from 'lib-components/atoms/buttons/InlineButton'
 import { desktopMin } from 'lib-components/breakpoints'
-import { isGroupMessageAccount } from 'lib-components/employee/messages/types'
 import { FixedSpaceColumn } from 'lib-components/layout/flex-helpers'
+import { isGroupMessageAccount } from 'lib-components/messages/types'
 import { fontWeights, NavLinkText } from 'lib-components/typography'
 import { BaseProps } from 'lib-components/utils'
 import { defaultMargins } from 'lib-components/white-space'
