+++ frontend/src/employee-mobile-frontend/messages/MessagesPage.tsx
@@ -9,8 +9,8 @@ import styled from 'styled-components'
 import { GroupInfo } from 'lib-common/generated/api-types/attendance'
 import { UUID } from 'lib-common/types'
 import useNonNullableParams from 'lib-common/useNonNullableParams'
-import EmptyMessageFolder from 'lib-components/employee/messages/EmptyMessageFolder'
 import { ContentArea } from 'lib-components/layout/Container'
+import EmptyMessageFolder from 'lib-components/messages/EmptyMessageFolder'
 import { H1 } from 'lib-components/typography'
 import { defaultMargins } from 'lib-components/white-space'
 import colors from 'lib-customizations/common'
