+++ frontend/src/employee-mobile-frontend/messages/api.ts
@@ -18,7 +18,7 @@ import {
 } from 'lib-common/generated/api-types/messaging'
 import { JsonOf } from 'lib-common/json'
 import { UUID } from 'lib-common/types'
-import { SaveDraftParams } from 'lib-components/employee/messages/types'
+import { SaveDraftParams } from 'lib-components/messages/types'
 
 import { API_URL, client } from '../client'
